Greatly improved parse error reporting
Now only the parse errors are reported that have been found at the
"deepest" level it was able to parse. This may still involve multiple
errors, as is expected.

Moreover, it now retains errors found while trying to parse optional
parts of the rules. This will improve the usability as well as simplify
the support for the ?, + and * modifiers.