Have cuts as proper parsers, instead of keyword
The upside is a cleaner implementation and less chance of a typo.

It would also clear the way, for if we would decide that passing
keyword as parsers would be a good idea. It might, it might not.

The downside is that the parsers can now be used anywhere, instead of
nicely restricted to chains, at the right places.