Performance tuning through caching only certain tree depth
This adds an extra attribute to success nodes: its depth. It counts
how many layers of children a node has. Using this, the caching
strategy can be altered such that only nodes of a particular depth are
cached. This can come at the cost of speed, but can significantly
reduce memory pressure.

That said, using hard-cuts should still be the main strategy of
reducing memory usage.