Grammar combinator now returns a parser
Before this change, the grammar macro would return a map. This could
lead to the impression that this map could be reused or altered, and
the refs would update. But this is a false impression. The refs are
bound as soon as the (outer) grammar combinator is done, and cannot be
rebound.

This change does lead to forcing a :root rule in the grammar. If a
grammar does not have this, it is easily added, since the grammar
combinator takes multiple maps. The same strategy can be used when
testing individual rules (i.e. starting the parse from a different
rule than root).