Hiccup parser model (#24)
This concerns a refactor, where every grammar (string-based, data-based and the new combinator-based) now return a vector model in "hiccup"-style. The `core/parse` function will compile this on the fly (using the vector-grammar), or it can be compiled beforehand using the new `core/compile` function.

This allows us to get rid of the `grammar` macro, and opens up the easy possibility to combine grammars (something that was a bit cumbersome). It has been replaced with a `with-scope` macro, that now even supports lexical scoping.

See release or PR details for details.